refactor: constants 폴더 구조 수정
