+++ frontend/components/filter/experienceLevelFilter/index.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 
-import { PARAMS } from '@/constants/api/queryParams';
-import { EXPERIENCE_LEVEL_RANGE, generateExperienceLevelText } from '@/constants/components/experienceLevel';
+import { PARAMS } from '@/constants/api';
+import { EXPERIENCE_LEVEL_RANGE, generateExperienceLevelText } from '@/constants/components';
 
 import { Slider, Text } from '@/components/common';
