+++ frontend/components/filter/jobsFilter/components/applyJobsButton.tsx
@@ -1,6 +1,6 @@
 import { Job } from '@/types/api/jobs';
 
-import { PARAMS } from '@/constants/api/queryParams';
+import { PARAMS } from '@/constants/api';
 
 import { Button, Text } from '@/components/common';
