+++ frontend/components/filter/serviceStatusFilter/index.tsx
@@ -1,4 +1,4 @@
-import { PARAMS, SERVICE_STATUSES } from '@/constants/api/queryParams';
+import { PARAMS, SERVICE_STATUSES } from '@/constants/api';
 
 import { Radio, Text } from '@/components/common';
