+++ frontend/constants/api/sort.ts
@@ -1,0 +1,1 @@
+export const SORT_TYPES = ['최신순', '마감순'] as const;
