+++ frontend/constants/components/index.ts
@@ -1,0 +1,1 @@
+export { EXPERIENCE_LEVEL_RANGE, generateExperienceLevelText } from './experienceLevel';
