--- frontend/constants/components/sort.ts
@@ -1,1 +1,0 @@
-export const SORT_TYPES = ['최신순', '마감순'];
