+++ frontend/hooks/useEducationLevel.ts
@@ -2,7 +2,7 @@ import { useSearchParams } from 'next/navigation';
 
 import { isValidEducationLevel } from '@/types/guards/queryParams';
 
-import { PARAMS } from '@/constants/api/queryParams';
+import { PARAMS } from '@/constants/api';
 
 export const useEducationLevel = () => {
   const searchParams = useSearchParams();
