+++ frontend/hooks/useLocations.ts
@@ -2,7 +2,7 @@ import { useSearchParams } from 'next/navigation';
 
 import { isValidLocations } from '@/types/guards/queryParams';
 
-import { PARAMS } from '@/constants/api/queryParams';
+import { PARAMS } from '@/constants/api';
 
 export const useLocations = () => {
   const searchParams = useSearchParams();
