+++ frontend/types/api/educationLevel.ts
@@ -1,3 +1,3 @@
-import { EDUCATION_LEVELS } from '@/constants/api/educationLevel';
+import { EDUCATION_LEVELS } from '@/constants/api';
 
 export type EducationLevel = (typeof EDUCATION_LEVELS)[number];
